Grocery fix

+++ app/src/main/java/com/github/houseorganizer/houseorganizer/panels/household/EditHouseholdActivity.java
@@ -325,7 +325,8 @@ public class EditHouseholdActivity extends ThemedAppCompatActivity {
         return firestore.collection("shop_lists")
                 .whereEqualTo("household", currentHousehold)
                 .get().addOnCompleteListener(doc ->{
-                    assert(doc.getResult().size() == 1);
+                    assert(doc.getResult().size() <= 1);
+                    if(doc.getResult().size() < 1)return;
                     doc.getResult().getDocuments().get(0).getReference().delete();
                 }).addOnFailureListener(toastExceptionFailureListener("Could not delete groceries"));
     }
